Move homepage to its own file
