+++ webapi/templates/homepage.html
@@ -48,6 +48,7 @@
             <tr><td>Network:</td><td>{{ .Network }}</td></tr>
             <tr><td>Support:</td><td>{{ .SupportEmail }}</td></tr>
             <tr><td>Pubkey:</td><td>{{ printf "%x" .PubKey }}</td></tr>
+            <tr><td>VSP closed:</td><td>{{ .VspClosed }}</td></tr>
         </table>
         <p>Last updated: {{.UpdateTime}}</p>
     </body>
